+++ wideq/core.py
@@ -1,11 +1,12 @@
 """A low-level, general abstraction for the LG SmartThinQ API.
 """
-import ssl
 import requests
 from urllib3.poolmanager import PoolManager
 from requests.adapters import HTTPAdapter
-from urllib.parse import urljoin, urlencode, urlparse, parse_qs
+import ssl
 import base64
+import uuid
+from urllib.parse import urljoin, urlencode, urlparse, parse_qs
 import hashlib
 import hmac
 import logging
@@ -14,7 +15,43 @@ from typing import Any, Dict, Generator, Optional
 
 from . import as_list, gen_uuid
 from . import core_exceptions as exc
-from .device import DeviceInfo, DEFAULT_TIMEOUT, DEFAULT_REFRESH_TIMEOUT
+from .device import DeviceInfo, DEFAULT_TIMEOUT
+
+import os
+import json
+
+DEFAULT_COUNTRY = "US"
+DEFAULT_LANGUAGE = "en-US"
+
+# v2
+V2_API_KEY = "VGhpblEyLjAgU0VSVklDRQ=="
+V2_CLIENT_ID = "65260af7e8e6547b51fdccf930097c51eb9885a508d3fddfa9ee6cdec22ae1bd"
+V2_MESSAGE_ID = "wideq"
+V2_SVC_PHASE = "OP"
+V2_APP_LEVEL = "PRD"
+V2_APP_OS = "LINUX"
+V2_APP_TYPE = "NUTS"
+V2_APP_VER = "3.0.1700"
+
+# new
+V2_GATEWAY_URL = "https://route.lgthinq.com:46030/v1/service/application/gateway-uri"
+OAUTH_REDIRECT_URI = "https://kr.m.lgaccount.com/login/iabClose"
+
+# orig
+SECURITY_KEY = "nuts_securitykey"
+SVC_CODE = "SVC202"
+DATA_ROOT = "lgedmRoot"
+CLIENT_ID = "LGAO221A02"
+OAUTH_SECRET_KEY = "c053c2a6ddeb7ad97cb0eed0dcb31cf8"
+DATE_FORMAT = "%a, %d %b %Y %H:%M:%S +0000"
+
+API2_ERRORS = {
+    "0102": exc.NotLoggedInError,
+    "0106": exc.NotConnectedError,
+}
+
+MIN_TIME_BETWEEN_UPDATE = 25  # seconds
+_LOGGER = logging.getLogger(__name__)
 
 
 class Tlsv1HttpAdapter(HTTPAdapter):
@@ -25,29 +62,6 @@ class Tlsv1HttpAdapter(HTTPAdapter):
             block=block,
             ssl_version=ssl.PROTOCOL_TLSv1,
         )
-
-
-GATEWAY_URL = "https://kic.lgthinq.com:46030/api/common/gatewayUriList"
-APP_KEY = "wideq"
-SECURITY_KEY = "nuts_securitykey"
-DATA_ROOT = "lgedmRoot"
-SVC_CODE = "SVC202"
-CLIENT_ID = "LGAO221A02"
-OAUTH_SECRET_KEY = "c053c2a6ddeb7ad97cb0eed0dcb31cf8"
-OAUTH_CLIENT_KEY = "LGAO221A02"
-DATE_FORMAT = "%a, %d %b %Y %H:%M:%S +0000"
-
-DEFAULT_COUNTRY = "US"
-DEFAULT_LANGUAGE = "en-US"
-
-API_ERRORS = {
-    "0102": exc.NotLoggedInError,
-    "0106": exc.NotConnectedError,
-    "0100": exc.FailedRequestError,
-    9000: exc.InvalidRequestError,  # Surprisingly, an integer (not a string).
-}
-
-_LOGGER = logging.getLogger(__name__)
 
 
 def oauth2_signature(message, secret):
@@ -64,63 +78,52 @@ def oauth2_signature(message, secret):
     return base64.b64encode(digest)
 
 
-def lgedm_post(url, data=None, access_token=None, session_id=None, use_tlsv1=True):
-    """Make an HTTP request in the format used by the API servers.
-
-    In this format, the request POST data sent as JSON under a special
-    key; authentication sent in headers. Return the JSON data extracted
-    from the response.
-
-    The `access_token` and `session_id` are required for most normal,
-    authenticated requests. They are not required, for example, to load
-    the gateway server data or to start a session.
-    """
-
-    _LOGGER.debug("lgedm_post before: %s", url)
+def thinq2_headers(
+    extra_headers={},
+    access_token=None,
+    user_number=None,
+    country=DEFAULT_COUNTRY,
+    language=DEFAULT_LANGUAGE,
+):
+    """Prepare API2 header."""
 
     headers = {
-        "x-thinq-application-key": APP_KEY,
-        "x-thinq-security-key": SECURITY_KEY,
         "Accept": "application/json",
+        "Content-type": "application/json;charset=UTF-8",
+        "x-api-key": V2_API_KEY,
+        "x-client-id": V2_CLIENT_ID,
+        "x-country-code": country,
+        "x-language-code": language,
+        "x-message-id": V2_MESSAGE_ID,
+        "x-service-code": SVC_CODE,
+        "x-service-phase": V2_SVC_PHASE,
+        "x-thinq-app-level": V2_APP_LEVEL,
+        "x-thinq-app-os": V2_APP_OS,
+        "x-thinq-app-type": V2_APP_TYPE,
+        "x-thinq-app-ver": V2_APP_VER,
+        "x-thinq-security-key": SECURITY_KEY,
     }
+
     if access_token:
-        headers["x-thinq-token"] = access_token
-    if session_id:
-        headers["x-thinq-jsessionId"] = session_id
+        headers["x-emp-token"] = access_token
 
-    s = requests.Session()
-    if use_tlsv1:
-        s.mount(url, Tlsv1HttpAdapter())
-    res = s.post(url, json={DATA_ROOT: data}, headers=headers, timeout=DEFAULT_TIMEOUT)
-    # res = requests.post(url, json={DATA_ROOT: data}, headers=headers, timeout = DEFAULT_TIMEOUT)
+    if user_number:
+        headers["x-user-no"] = user_number
 
-    out = res.json()
-    _LOGGER.debug("lgedm_post after: %s", out)
-
-    msg = out.get(DATA_ROOT)
-    if not msg:
-        raise exc.APIError("-1", out)
-
-    # Check for API errors.
-    if "returnCd" in msg:
-        code = msg["returnCd"]
-        if code != "0000":
-            message = msg["returnMsg"]
-            if code in API_ERRORS:
-                raise API_ERRORS[code]()
-            raise exc.APIError(code, message)
-
-    return msg
+    return {**headers, **extra_headers}
 
 
-def gateway_info(country, language):
-    """Load information about the hosts to use for API interaction.
+def thinq2_get(
+    url,
+    access_token=None,
+    user_number=None,
+    headers={},
+    country=DEFAULT_COUNTRY,
+    language=DEFAULT_LANGUAGE,
+):
+    """Make an HTTP request in the format used by the API2 servers."""
 
-    `country` and `language` are codes, like "US" and "en-US,"
-    respectively.
-    """
-
-    # this code to avoid ssl error with DH
+    # this code to avoid ssl error 'dh key too small'
     requests.packages.urllib3.disable_warnings()
     requests.packages.urllib3.util.ssl_.DEFAULT_CIPHERS += "HIGH:!DH:!aNULL"
     try:
@@ -130,29 +133,88 @@ def gateway_info(country, language):
     except AttributeError:
         # no pyopenssl support used / needed / available
         pass
-    # this code to avoid ssl error with DH
+    # this code to avoid ssl error 'dh key too small'
 
-    return lgedm_post(GATEWAY_URL, {"countryCode": country, "langCode": language},)
+    _LOGGER.debug("thinq2_get before: %s", url)
 
-
-def oauth_url(auth_base, country, language):
-    """Construct the URL for users to log in (in a browser) to start an
-    authenticated session.
-    """
-
-    url = urljoin(auth_base, "login/sign_in")
-    query = urlencode(
-        {
-            "country": country,
-            "language": language,
-            "svcCode": SVC_CODE,
-            "authSvr": "oauth2",
-            "client_id": CLIENT_ID,
-            "division": "ha",
-            "grant_type": "password",
-        }
+    res = requests.get(
+        url,
+        headers=thinq2_headers(
+            access_token=access_token,
+            user_number=user_number,
+            extra_headers=headers,
+            country=country,
+            language=language,
+        ),
+        timeout=DEFAULT_TIMEOUT,
     )
-    return "{}?{}".format(url, query)
+
+    out = res.json()
+    _LOGGER.debug("thinq2_get after: %s", out)
+
+    if "resultCode" not in out:
+        raise exc.APIError("-1", out)
+
+    code = out["resultCode"]
+    if code != "0000":
+        if code in API2_ERRORS:
+            raise API2_ERRORS[code]()
+        raise exc.APIError(code, "error")
+    return out["result"]
+
+
+def lgedm2_post(
+    url,
+    data=None,
+    access_token=None,
+    user_number=None,
+    headers={},
+    country=DEFAULT_COUNTRY,
+    language=DEFAULT_LANGUAGE,
+    use_tlsv1=True,
+):
+    """Make an HTTP request in the format used by the API servers."""
+
+    _LOGGER.debug("lgedm2_post before: %s", url)
+
+    s = requests.Session()
+    if use_tlsv1:
+        s.mount(url, Tlsv1HttpAdapter())
+    res = s.post(
+        url,
+        json={DATA_ROOT: data},
+        headers=thinq2_headers(
+            access_token=access_token,
+            user_number=user_number,
+            extra_headers=headers,
+            country=country,
+            language=language,
+        ),
+        timeout=DEFAULT_TIMEOUT,
+    )
+
+    out = res.json()
+    _LOGGER.debug("lgedm2_post after: %s", out)
+
+    msg = out.get(DATA_ROOT)
+    if not msg:
+        raise exc.APIError("-1", out)
+
+    if "returnCd" in msg:
+        code = msg["returnCd"]
+        if code != "0000":
+            message = msg["returnMsg"]
+            if code in API2_ERRORS:
+                raise API2_ERRORS[code]()
+            raise exc.APIError(code, message)
+
+    return msg
+
+
+def gateway_info(country, language):
+    """ TODO
+    """
+    return thinq2_get(V2_GATEWAY_URL, country=country, language=language)
 
 
 def parse_oauth_callback(url):
@@ -162,163 +224,217 @@ def parse_oauth_callback(url):
     """
 
     params = parse_qs(urlparse(url).query)
-    return params["access_token"][0], params["refresh_token"][0]
+    return params["oauth2_backend_url"][0], params["code"][0], params["user_number"][0]
 
 
-def login(api_root, access_token, country, language):
-    """Use an access token to log into the API and obtain a session and
-    return information about the session.
+def auth_request(oauth_url, data):
+    """Use an auth code to log into the v2 API and obtain an access token 
+    and refresh token.
     """
-
-    url = urljoin(api_root + "/", "member/login")
-    data = {
-        "countryCode": country,
-        "langCode": language,
-        "loginType": "EMP",
-        "token": access_token,
-    }
-    return lgedm_post(url, data)
-
-
-def refresh_auth(oauth_root, refresh_token, use_tlsv1=True):
-    """Get a new access_token using a refresh_token.
-
-    May raise a `TokenError`.
-    """
-
-    token_url = urljoin(oauth_root, "/oauth2/token")
-    data = {
-        "grant_type": "refresh_token",
-        "refresh_token": refresh_token,
-    }
-
-    # The timestamp for labeling OAuth requests can be obtained
-    # through a request to the date/time endpoint:
-    # https://us.lgeapi.com/datetime
-    # But we can also just generate a timestamp.
+    auth_path = "/oauth/1.0/oauth2/token"
+    url = urljoin(oauth_url, "/oauth/1.0/oauth2/token")
     timestamp = datetime.utcnow().strftime(DATE_FORMAT)
-
-    # The signature for the requests is on a string consisting of two
-    # parts: (1) a fake request URL containing the refresh token, and (2)
-    # the timestamp.
-    req_url = "/oauth2/token?grant_type=refresh_token&refresh_token=" + refresh_token
+    req_url = "{}?{}".format(auth_path, urlencode(data))
     sig = oauth2_signature("{}\n{}".format(req_url, timestamp), OAUTH_SECRET_KEY)
 
     headers = {
-        "lgemp-x-app-key": OAUTH_CLIENT_KEY,
-        "lgemp-x-signature": sig,
-        "lgemp-x-date": timestamp,
+        "x-lge-appkey": CLIENT_ID,
+        "x-lge-oauth-signature": sig,
+        "x-lge-oauth-date": timestamp,
         "Accept": "application/json",
     }
 
-    s = requests.Session()
-    if use_tlsv1:
-        s.mount(token_url, Tlsv1HttpAdapter())
-    res = s.post(token_url, data=data, headers=headers, timeout=DEFAULT_REFRESH_TIMEOUT)
-    # res = requests.post(token_url, data=data, headers=headers, timeout = DEFAULT_REFRESH_TIMEOUT)
+    res = requests.post(url, headers=headers, data=data, timeout=DEFAULT_TIMEOUT)
+
+    if res.status_code != 200:
+        raise exc.TokenError()
 
     res_data = res.json()
     _LOGGER.debug(res_data)
 
-    if res_data["status"] != 1:
-        raise exc.TokenError()
-    return res_data["access_token"]
+    return res_data
+
+
+def login(oauth_url, auth_code):
+    """Get a new access_token using an authorization_code
+    
+    May raise a `tokenError`.
+    """
+
+    out = auth_request(
+        oauth_url,
+        {
+            "code": auth_code,
+            "grant_type": "authorization_code",
+            "redirect_uri": OAUTH_REDIRECT_URI,
+        },
+    )
+
+    return out["access_token"], out["refresh_token"]
+
+
+def refresh_auth(oauth_root, refresh_token):
+    """Get a new access_token using a refresh_token.
+
+    May raise a `TokenError`.
+    """
+    out = auth_request(
+        oauth_root, {"grant_type": "refresh_token", "refresh_token": refresh_token}
+    )
+
+    return out["access_token"]
 
 
 class Gateway(object):
-    def __init__(self, auth_base, api_root, oauth_root, country, language):
+    def __init__(self, auth_base, api_root, api2_root, country, language):
         self.auth_base = auth_base
         self.api_root = api_root
-        self.oauth_root = oauth_root
+        self.api2_root = api2_root
         self.country = country
         self.language = language
 
     @classmethod
     def discover(cls, country, language):
         gw = gateway_info(country, language)
-        return cls(gw["empUri"], gw["thinqUri"], gw["oauthUri"], country, language)
-
-    def get_tokens(self, url):
-        """Create an authentication using an OAuth callback URL.
-        """
-        access_token, refresh_token = parse_oauth_callback(url)
-        return {"access_token": access_token, "refresh_token": refresh_token}
+        return cls(gw["empUri"], gw["thinq1Uri"], gw["thinq2Uri"], country, language)
 
     def oauth_url(self):
-        return oauth_url(self.auth_base, self.country, self.language)
+        """Construct the URL for users to log in (in a browser) to start an
+        authenticated session.
+        """
+
+        url = urljoin(self.auth_base, "spx/login/signIn")
+        query = urlencode(
+            {
+                "country": self.country,
+                "language": self.language,
+                "svc_list": SVC_CODE,
+                "client_id": CLIENT_ID,
+                "division": "ha",
+                "redirect_uri": OAUTH_REDIRECT_URI,
+                "state": uuid.uuid1().hex,
+                "show_thirdparty_login": "GGL,AMZ,FBK",
+            }
+        )
+        return "{}?{}".format(url, query)
 
     def dump(self):
         return {
             "auth_base": self.auth_base,
             "api_root": self.api_root,
-            "oauth_root": self.oauth_root,
+            "api2_root": self.api2_root,
             "country": self.country,
             "language": self.language,
         }
 
 
 class Auth(object):
-    def __init__(self, gateway, access_token, refresh_token):
+    def __init__(self, gateway, oauth_url, access_token, refresh_token, user_number):
         self.gateway = gateway
+        self.oauth_url = oauth_url
         self.access_token = access_token
         self.refresh_token = refresh_token
+        self.user_number = user_number
 
     @classmethod
     def from_url(cls, gateway, url):
         """Create an authentication using an OAuth callback URL.
         """
+        oauth_url, auth_code, user_number = parse_oauth_callback(url)
+        access_token, refresh_token = login(oauth_url, auth_code)
 
-        access_token, refresh_token = parse_oauth_callback(url)
-        return cls(gateway, access_token, refresh_token)
+        return cls(gateway, oauth_url, access_token, refresh_token, user_number)
 
     def start_session(self):
         """Start an API session for the logged-in user. Return the
         Session object and a list of the user's devices.
         """
-
-        session_info = login(
-            self.gateway.api_root,
-            self.access_token,
-            self.gateway.country,
-            self.gateway.language,
-        )
-        session_id = session_info["jsessionId"]
-        devices = session_info.get("item", [])
-        return Session(self, session_id), as_list(devices)
+        return Session(self)
 
     def refresh(self):
         """Refresh the authentication, returning a new Auth object.
         """
 
-        new_access_token = refresh_auth(self.gateway.oauth_root, self.refresh_token)
-        return Auth(self.gateway, new_access_token, self.refresh_token)
+        new_access_token = refresh_auth(self.oauth_url, self.refresh_token)
+        return Auth(
+            self.gateway,
+            self.oauth_url,
+            new_access_token,
+            self.refresh_token,
+            self.user_number,
+        )
 
     def dump(self):
-        return {"access_token": self.access_token, "refresh_token": self.refresh_token}
+        return {
+            "access_token": self.access_token,
+            "refresh_token": self.refresh_token,
+            "oauth_url": self.oauth_url,
+            "user_number": self.user_number,
+        }
+
+    @classmethod
+    def load(cls, gateway, data):
+        return cls(
+            gateway,
+            data["oauth_url"],
+            data["access_token"],
+            data["refresh_token"],
+            data["user_number"],
+        )
 
 
 class Session(object):
-    def __init__(self, auth, session_id):
+    def __init__(self, auth, session_id=None):
         self.auth = auth
         self.session_id = session_id
 
     def post(self, path, data=None):
-        """Make a POST request to the API server.
+        """Make a POST request to the APIv1 server.
 
         This is like `lgedm_post`, but it pulls the context for the
         request from an active Session.
         """
 
         url = urljoin(self.auth.gateway.api_root + "/", path)
-        return lgedm_post(url, data, self.auth.access_token, self.session_id)
+        return lgedm2_post(
+            url,
+            data,
+            self.auth.access_token,
+            self.auth.user_number,
+            country=self.auth.gateway.country,
+            language=self.auth.gateway.language,
+        )
+
+    def get(self, path):
+        """Make a GET request to the APIv1 server."""
+
+        url = urljoin(self.auth.gateway.api_root + "/", path)
+        return thinq2_get(
+            url,
+            self.auth.access_token,
+            self.auth.user_number,
+            country=self.auth.gateway.country,
+            language=self.auth.gateway.language,
+        )
+
+    def get2(self, path):
+        """Make a GET request to the APIv2 server."""
+
+        url = urljoin(self.auth.gateway.api2_root + "/", path)
+        return thinq2_get(
+            url,
+            self.auth.access_token,
+            self.auth.user_number,
+            country=self.auth.gateway.country,
+            language=self.auth.gateway.language,
+        )
 
     def get_devices(self):
         """Get a list of devices associated with the user's account.
 
         Return a list of dicts with information about the devices.
         """
-
-        devices = self.post("device/deviceList").get("item", [])
+        devices = self.get2("service/application/dashboard").get("item", [])
         return as_list(devices)
 
     def monitor_start(self, device_id):
@@ -441,6 +557,7 @@ class Client(object):
         self._gateway: Optional[Gateway] = gateway
         self._auth: Optional[Auth] = auth
         self._session: Optional[Session] = session
+        self._last_device_update = datetime.now()
 
         # The last list of devices we got from the server. This is the
         # raw JSON list data describing the devices.
@@ -453,6 +570,22 @@ class Client(object):
         # Locale information used to discover a gateway, if necessary.
         self._country = country
         self._language = language
+
+    def _inject_thinq2_device(self):
+        """This is used only for debug"""
+        data_file = os.path.dirname(os.path.realpath(__file__)) + "/deviceV2.txt"
+        with open(data_file, "r") as f:
+            device_v2 = json.load(f)
+        for d in device_v2:
+            self._devices.append(d)
+            _LOGGER.debug("Injected debug device: %s", d)
+
+    def _load_devices(self, force_update: bool = False):
+        if self._session and (self._devices is None or force_update):
+            self._devices = self._session.get_devices()
+            # for debug
+            # self._inject_thinq2_device()
+            # for debug
 
     @property
     def gateway(self) -> Gateway:
@@ -469,7 +602,8 @@ class Client(object):
     @property
     def session(self) -> Session:
         if not self._session:
-            self._session, self._devices = self.auth.start_session()
+            self._session = self.auth.start_session()
+            self._load_devices()
         return self._session
 
     @property
@@ -480,20 +614,22 @@ class Client(object):
     def devices(self) -> Generator["DeviceInfo", None, None]:
         """DeviceInfo objects describing the user's devices.
             """
-
         if self._devices is None:
-            self._devices = self.session.get_devices()
+            self._load_devices()
         return (DeviceInfo(d) for d in self._devices)
 
     def refresh_devices(self):
-        return
+        call_time = datetime.now()
+        difference = (call_time - self._last_device_update).total_seconds()
+        if difference > MIN_TIME_BETWEEN_UPDATE:
+            self._load_devices(True)
+            self._last_device_update = call_time
 
     def get_device(self, device_id) -> Optional["DeviceInfo"]:
         """Look up a DeviceInfo object by device ID.
             
-            Return None if the device does not exist.
-            """
-
+        Return None if the device does not exist.
+        """
         for device in self.devices:
             if device.id == device_id:
                 return device
@@ -511,16 +647,14 @@ class Client(object):
             client._gateway = Gateway(
                 data["auth_base"],
                 data["api_root"],
-                data["oauth_root"],
+                data["api2_root"],
                 data.get("country", DEFAULT_COUNTRY),
                 data.get("language", DEFAULT_LANGUAGE),
             )
 
         if "auth" in state:
             data = state["auth"]
-            client._auth = Auth(
-                client.gateway, data["access_token"], data["refresh_token"]
-            )
+            client._auth = Auth.load(client._gateway, data)
 
         if "session" in state:
             client._session = Session(client.auth, state["session"])
@@ -547,7 +681,7 @@ class Client(object):
             out["gateway"] = {
                 "auth_base": self._gateway.auth_base,
                 "api_root": self._gateway.api_root,
-                "oauth_root": self._gateway.oauth_root,
+                "api2_root": self._gateway.api2_root,
                 "country": self._gateway.country,
                 "language": self._gateway.language,
             }
@@ -568,10 +702,14 @@ class Client(object):
 
     def refresh(self) -> None:
         self._auth = self.auth.refresh()
-        self._session, self._devices = self.auth.start_session()
+        self._session = self.auth.start_session()
+        # self._device = None
+        self._load_devices()
 
     @classmethod
-    def from_token(cls, refresh_token, country=None, language=None) -> "Client":
+    def from_token(
+        cls, oauth_url, refresh_token, user_number, country=None, language=None
+    ) -> "Client":
         """Construct a client using just a refresh token.
             
             This allows simpler state storage (e.g., for human-written
@@ -580,9 +718,10 @@ class Client(object):
             """
 
         client = cls(
-            country=country or DEFAULT_COUNTRY, language=language or DEFAULT_LANGUAGE,
+            country=country or DEFAULT_COUNTRY,
+            language=language or DEFAULT_LANGUAGE,
         )
-        client._auth = Auth(client.gateway, None, refresh_token)
+        client._auth = Auth(client.gateway, oauth_url, None, refresh_token, user_number)
         client.refresh()
         return client
 
@@ -590,8 +729,15 @@ class Client(object):
     def oauthinfo_from_url(cls, url):
         """Create an authentication using an OAuth callback URL.
         """
-        access_token, refresh_token = parse_oauth_callback(url)
-        return {"access_token": access_token, "refresh_token": refresh_token}
+        oauth_url, auth_code, user_number = parse_oauth_callback(url)
+        access_token, refresh_token = login(oauth_url, auth_code)
+
+        return {
+            "oauth_url": oauth_url,
+            "user_number": user_number,
+            "access_token": access_token,
+            "refresh_token": refresh_token,
+        }
 
     def model_info(self, device):
         """For a DeviceInfo object, get a ModelInfo object describing
